Add files via upload
Installation of game wise design

This application requires a sound library which is not included in this repository. Size of the audio library is around 1.3 gb which contains bin-aural recordings and sound design files.

Hardware:
A bicycle (also a stand for back wheel to stand it up)
Arduino Uno
2X GY 521 board (MPU6050 chip with accelerometer and gyro which have 6DOF) (one for helmet, one for handlebar)
A magnetic speed sensor
A projector to screen visuals

Software:
Max 6.1
Processing 2.2.1 with Video and OSC libraries
Arduino Software with i2cdevlib library for extracting DMP data from GY 521 board

Explanations:

This is a horror type of game project that aims more experiencing what it provides than achieving a goal as typical in simulants. It obtains a sound design work which leads the user while interacting with it and a visual design that creates a scary atmosphere rather than deriving instructions or routing from the screen for users. Interactivity is achieved